Move ReturnTypeInfo to type_info.h
